fix: deep QA review — 30+ issues fixed across AI engine
7-agent code review found 90+ issues (50 unique). Fixed all critical+high:

DB (migration #16):
- Add UPDATE RLS policy on ai_research_cache (upsert requires it)
- Remove redundant index (duplicated UNIQUE constraint)
- Add NOT NULL on created_at, model_used, prompt_version

Retry/escalation (3 bugs):
- withRetry now skips non-transient errors (AIValidationError, CONFIG_ERROR)
- withModelEscalation only escalates transient errors + logs fallback
- Remove identical-endpoint escalation in market-insights synthesis
- Remove MODEL_ESCALATION dead code, fix unreachable lastError

Silent failures (8 fixes):
- All cache upserts now check { error } and log failures
- Cache reads check error (distinguish DB failure from cache miss)
- searchWebParallel throws on ALL queries failing (was silent [])
- extractFromSources logs individual extraction failures
- Deep research warns on zero search results

API routes (12 fixes):
- Add .max() on all string inputs (prevent multi-MB payloads)
- Add POST handler for deep research trigger (GET kept for compat)
- Validate cache_key as 64-char hex SHA-256
- Validate action query param via Zod enum
- Use AI_CONFIG model name instead of hardcoded string
- Validate search_params from DB before using (no raw 'as' cast)
- Add .int() on feedback score, .min(1) on category
- Log req.json() parse errors

Security:
- Add sanitizeInput() for user strings in AI prompts
- Apply to all 4 prompt template files (market, JD, stage, feedback)

Schemas:
- Add salary min<=max refinement
- Add .int() on FocusArea weight
- Add .min(1).max(4) bounds on average_score

Types:
- Align domain types with Zod schemas (JobDescription.requirements,
  FocusArea.name, SuggestedQuestion fields)
- Error code now typed union (AIErrorCode)
- Error classes support { cause } for stack trace preservation

Tests: 77/77 passing (added 4 new retry/escalation tests)

+++ packages/ai/src/schemas/interview-stage.ts
@@ -3,7 +3,7 @@ import { z } from "zod";
 export const FocusAreaSchema = z.object({
   name: z.string(),
   description: z.string(),
-  weight: z.number().min(1).max(4),
+  weight: z.number().int().min(1).max(4),
 });
 
 export const SuggestedQuestionSchema = z.object({
